- Changed the Limit parameter type from byt to int. - Added tests for the GetCategories endpoint.

+++ src/Plugins/Nop.Plugin.Api/Models/CategoriesParameters/CategoriesParametersModel.cs
@@ -18,7 +18,7 @@ namespace Nop.Plugin.Api.Models.CategoriesParameters
 
         public string Ids { get; set; }
 
-        public byte Limit { get; set; }
+        public int Limit { get; set; }
 
         public int Page { get; set; }
